Make BasePackedGraph not inherit any virtual anything and proxy around it to add dynamic dispatch
